+++ SOUP.user.js
@@ -1,9 +1,9 @@
 // ==UserScript==
 // @name        Stack Overflow Unofficial Patch
 // @namespace   https://github.com/vyznev/
-// @description Miscellaneous client-side fixes for bugs on Stack Exchange sites
+// @description Miscellaneous client-side fixes for bugs on Stack Exchange sites (development)
 // @author      Ilmari Karonen
-// @version     1.22.0
+// @version     1.23.0
 // @copyright   2014, Ilmari Karonen (http://stackapps.com/users/10283/ilmari-karonen)
 // @license     ISC; http://opensource.org/licenses/ISC
 // @match       *://*.stackexchange.com/*
@@ -14,9 +14,9 @@
 // @match       *://*.mathoverflow.net/*
 // @match       *://*.askubuntu.com/*
 // @homepageURL http://stackapps.com/questions/4486/stack-overflow-unofficial-patch
-// @updateURL   https://github.com/vyznev/soup/raw/master/SOUP.meta.js
-// @downloadURL https://github.com/vyznev/soup/raw/master/SOUP.user.js
-// @icon        https://github.com/vyznev/soup/raw/master/icon/SOUP_icon_128.png
+// @updateURL   https://github.com/vyznev/soup/raw/devel/SOUP.meta.js
+// @downloadURL https://github.com/vyznev/soup/raw/devel/SOUP.user.js
+// @icon        https://github.com/vyznev/soup/raw/devel/icon/SOUP_icon_128.png
 // @grant       none
 // @run-at      document-start
 // ==/UserScript==
@@ -731,6 +731,39 @@ fixes.mse266852 = {
 			$('div[id^="comments-link-"] .js-link-separator:not(.lsep)').addClass('lsep').text('|');
 		} ).code();
 	}
+};
+fixes.mse239382 = {
+	title:	"Earned bounties layout is broken",
+	url:	"http://meta.stackexchange.com/q/239382",
+	script:	function () {
+		// fix #1: wrap header in <h1> tags
+		SOUP.hookAjax( /^\/ajax\/users\/panel\//, function () {
+			$('body.user-page .user-panel .subheader > a').wrap('<h1></h1>');
+		} ).code();
+		
+		// fix #2: make the change event handler live
+		var selector = '.user-panel-subtabs select'; var matches = $(selector);
+		if ( ! matches.length ) return;
+		$._data( matches[0], 'events' ).change.forEach( function ( h ) {
+			if ( h.selector || ! /"div\[class='subheader'\]"/.test( h.handler.toString() ) ) return;
+			$('body').on( 'change', selector, h.handler );
+			matches.off( 'change', h.handler );
+		} );
+		
+		// fix #3: sync the selector with the visible content
+		// XXX: this will only work on English sites, and may break there if the title text is changed
+		matches.each( function () {
+			var title = $(this).closest('.user-panel').find('div.subheader').text().toLowerCase();
+			for (var i = 0; i < this.options.length; i++ ) {
+				if ( title.indexOf( this.options[i].value ) < 0 ) continue;
+				this.options[i].selected = true;
+			}
+		} )
+
+	},
+	// fix #4: clean up excess whitespace by making sorter overlap table
+	css:	".user-show-new .user-panel .user-panel-subtabs.sorter" +
+		" { margin-bottom: -100%; position: relative; background: white; padding: 1px }" 
 };